Use a vector of per-frame AliasClasses to represent backtrace memory effects
Summary:
Union of multiple ALocals corresponding to two or more frames is pessimized to
ALocalAny.

This pessimizes the backtrace component of CallEffects and GeneralEffects if
the translation has at least two frames with interesting locals, such as
metadata or coeffects. When that happens, store-elim is not able to sink writes
through these instructions and we end up spilling usually a ton of Uninits to
otherwise unused locals.

Reviewed By: ricklavoie

Differential Revision: D48812311

fbshipit-source-id: 9355ce93abd9c73cc51d2627684c83a429bafa83