Enable heterogeneous scope lookups
Summary: Enable heterogeneous lookups by `string_view` in scope. This will allow removing a  bunch of unnecessary `std::string` allocations and simplify code. Once we switch to C++20 we'll be able to use heterogeneous unordered maps but it's not particularly important since most of the time is spent in codegen and not here.

Reviewed By: avalonalex

Differential Revision: D48811305

fbshipit-source-id: 1eecbf45c0d6006533acf94d87dfdd2416abcb9a