Provide Reasons for Data Type
Summary:
# Context
When performing the well-formed check on case types that they do not have overlapping data types, we currently don't provide reasons why we believe a particular type can contain values with a certain runtime tag. When we add support for multiple object types it will be useful to provide more details in the error messages why we consider types to overlap and perhaps provide suggestions for how to work around the restriction we have in place.

Even without supporting multiple objects, it is useful to provide more details why some types are associated with a given runtime tag, even if it isn't obvious.

# Solution
This diff computes reason information when constructing nodes for our `DataType.Set`. This reason will be stored along with `Tag.t` value in the `DataType.Set` so we can provide more detailed error reporting. A reason consists of a subreason that explains special relationships, i.e. that `XHPChild` includes `string` because it a special interface that includes non-object members. It also has a trail, which traces the component types we navigate through to reach our determination of a tag. For instance, if a generic is included in case type we trace through its upper bounds and then this will show up as part of the trail. Finally there is an origin which traces back to the original variant type.

As part of this change I needed to change the interface for ApproxSet to allow for the Domain type to take a generic parameter so we can handle origins from different phases.

Reviewed By: enetsee

Differential Revision: D48653810

fbshipit-source-id: 2a0336ac5a86127a07cb86f0bf4b72bb9e5ac0df