Treat the None module as the default module when checking packages
Summary:
D48406136 makes it so that HHVM treats files with no module as the "default" module.

The only typechecker impact of this change is that when checking packages, we should treat modules as being part of the "default" module when globbing.

Regular module code isn't affected, so no reason to add a "default" string to every single TAST for now.

Differential Revision: D48527408

fbshipit-source-id: 168b87f5b8a624103d6393f85f6afc7a569bc81d