AllocationColocator
Summary:
One way to describe `AllocationColocator` is that it can define dynamically sized structs on the heap.

This class abstracts away the scary parts of managing colocated memory allocation and access — a pattern that's already used in `ContextStack`.

After this lands, we can use this abstraction in other parts of our codebase.

Reviewed By: iahs

Differential Revision: D48674477

fbshipit-source-id: e7d0de4428c89c6eec101e0b8a41a18e138a379d