Depgraph: allow to remove edges
Summary:
To do so, we add some set of removed edges to the depgraph delta, and adapt all the functions to account for them.
Before this diff, the struct for the delta was:
```
pub struct DepGraphDelta {
    /// Maps each dependency to a set of dependents
    rdeps: HashMap<Dep, HashSet<Dep>>,

    /// Total number of edges. Tracks the sum:
    /// `rdeps.values().map(|set|set.len()).sum()`
    num_edges: usize,
}
```
Now it's:
```
pub struct DepGraphDelta {
    /// Maps each dependency to a set of dependents
    added_edges: Graph,

    /// Edges to be removed from the base dependency graph.
    /// Invariant: `removed_edges` and `added_edges` are disjoint.
    removed_edges: Graph,
}
```
with
```
struct Graph {
    map: HashMap<Dep, HashSet<Dep>>,

    /// Total number of edges. Tracks the sum:
    /// `map.values().map(|set|set.len()).sum()`
    num_edges: usize,
}
```
Note that Graph is similar to the old DepGraphDelta.
As a consequence, lots of helpers on DepGraphDelta have been moved to Graph.

Instead of being adapted, some functions have been renamed to indicate that they only operate on the added edges, because I don't think they're used in a context where we're interested in edge removal, i.e. I believe they are only used when building the saved state from scratch using the whole of www with hh_distc. Please confirm or deny that assumption!

Differential Revision: D48575054

fbshipit-source-id: 2ff85cd96d0c187ea8ac880315a6dc2ba8d64cfe