find refs fix missing ide
Summary:
Max had observed that sometimes he was getting $/progress messages with partial results of find-refs, and sometimes not. Partly that was the race condition fixed in the previous diff. But also it was the oversight corrected here.

Design:
1. When the user does find-refs, we ask clientIdeDaemon for all references in all open files, plus an "action" by which we can ask hh_server for the remaining refs
2. We create a temporary stream_refs_pid.jsonl file into which hh_server will append its results. ClientLsp has an Lwt_stream.t through which it basically does "ptail -f" to read entries from that streaming file as they're written.
3. "hh --ide-find-refs-by-symbol3" causes hh_server to do its work, appending to the file as it goes
4. Whenever we find items in the Lwt_stream.t then we publish them as $/progress notifications to VSCode (so long as they don't relate to files already open in the IDE, for which clientIdeDaemon has already responded).
5. Once the shellout has finished, then we unlink the file, flush any remaining items from the file into the Lwt_stream.t, then close it, process all remaining Lwt_stream.t items into $/progress notifications, and finally send the LSP response.

Except we also want the initial clientIdeDaemon results (pertaining to open files) to get published as $/progress notifications as well. In my initial design, I'd achieved that by sticking them onto the newly created .jsonl file, so that the common subsequent codepath would read them from the .jsonl file in the normal way and place them on the Lwt_stream.t.

However there was a bug! Remember that when we read from the file, we filter out (decline to push to the stream) any results that pertain to IDE-open-files. Therefore, those initial items in the .jsonl file were never being pushed to the stream.

This diff fixes the bug! It makes it so the initial items get pushed straight to the Lwt_stream.t, bypassing the filter.

Reviewed By: nt591

Differential Revision: D48857306

fbshipit-source-id: e783c57d80a57e694e2f7794994a06e30d796bee