Merge constant into definition lookup
Summary:
It doesn't make sense to have separate maps for different kinds of top-level definitions because all of them share the same namespace (in the abstract sense, not a Thrift namespace). Now that we properly report redefinitions, merge the constant lookup into the general definition lookup. This has already been done for interactions and types in D48832714.

Get rid of the orgy of string and vector allocations on every constant.

Fix constness of `find` and `find_by_uri`.

Reviewed By: thedavekwon

Differential Revision: D48877675

fbshipit-source-id: 3aae9230d1084c54a1dbabdfc993f9713a9a167e