Thrift generated code changes for server streaming
Summary:
The handlers do the following:

First they call the client handler, and then get the first response type (if there is one), and the returned iterator.

The first response type is then serialized, and the iterator is passed into a wrapper that will get a value out of it, wrap it with the stream response type, serialize it, then return it.

Reviewed By: junzh0u

Differential Revision: D48691570

fbshipit-source-id: fecb93b358dba4e373c2d895bad382b5a12d3bea