dep_graph_has_edge_for_sure -> depgraph_has_edge
Summary: Instead of checking the base depgraph for the presence of and edge, which also required to check the removed edges in the delta, we simply check both the base and the delta (added + removed edges) for the difinite presence or absence of the edge. This way the semantic of the function is clearer.

Differential Revision: D48824893

fbshipit-source-id: 462f5f848769c3bb4e631431043471493bf13817