Fuse traversals
Summary:
When `--type-at-pos-batch` is used to query the typechecker for type information, we typecheck each file once, but traverse it as many times as there are arguments even if they all come from the same file.

This diff changes the `type_at_pos` traversal to handle multiple targets to the same TAST.

The following diffs in the stack incorporate this.

Reviewed By: enetsee

Differential Revision: D48933680

fbshipit-source-id: 9ba1bf04bc97ca5fc2204d629e51675b3d250f56