Remove tagging of declared members when detecting removed members
Summary:
The tagging of declared members in the dep graph needs to be maintained in sync with which members are actually declared.

We already mark members as declared while typechecking (D47471289), and we could "remove" the taggings at the same time. That would however add a lot of edges in the `removed_edges` field of the depgraph delta. One possible remediation is to query the presence of the tagging in the depgraph and only remove if present.

Another option is to rely on the member diffs computed just before computing fanouts to remove edges, which is what we do here: We compute diffs, compute the fanout based on these diffs, then remove the declared taggings based on those diffs.

Differential Revision: D48867370

fbshipit-source-id: 1fe215fa6061ff82b967b75c6d885212e7640e90