Add Support for Sealed Interfaces in Case Types
Summary:
In D48814014 we add the notion of an interface tag to our data type model. There isn't much we can conclude based on the premise of this tag. However we can take advantage of the sealed list on an interface to prove if types are disjoints. Consider the following:

```
case type CT = I1 | I2;

interface I1 {}

interface I2 {}
```

We would need to reject this case type since we cannot determine if the interfaces are disjoint. Now consider if we seal the interface `I1` like so:

```
<<__Sealed(C1::class, C2::class)>>
interface I1 {}

final class C1 implements I1 {}

final class C2 implements I1 {}
```

Based on the sealed list we can represent the data type of `I1` as `InstanceOf(C1, FinalClass) | InstanceOf(C2, FinalClass)` instead of `InstanceOf(I1, Interface)`. The data type is potentially larger than `InstanceOf(I1, Interface)` because we do not require classes in the sealed list to actually implement the interface, but it is always safe to approximate using a larger data type (for instance using `Tag.ObjectData` which is the set of all values with the object tag).

This can of course be applied recursively, so if `I1` was sealed to an interface and that interface had a sealed list then we can use that as part of determining the data type.

Reviewed By: vassilmladenov

Differential Revision: D48814039

fbshipit-source-id: c3c092a89ccf52162c20f082a7b989467effb9e5