Allow Multiple Final Class Types in Case Types
Summary:
This diff adds support for multiple class types in a case type if the classes are final. Since final classes cannot be extended, we know two final classes are disjoint if they have different names.

Supporting this required two small changes. First within the `Tag` module I add a new tag, `InstanceOf _` that is used to represent the tags associated with an object. Because of this in the `Tag.relation` we consider it to be a subset of the `ObjectData` tag.

The second change is within `DataType.Class.to_datatype`. When determining the data type for a class, we will look up the folded decl associated with the class. If it is a final class then we create the tag `InstanceOf (name, FinalClass)`.

Finally I updated the error reporting in `check_overlapping`, since after this change we expect to produce `Subset`/`Superset` relations.

Reviewed By: vassilmladenov

Differential Revision: D48768815

fbshipit-source-id: 840a4d3ec02afc331c848782f8d26a101dec73ef