Add Support for Interfaces in Case Types
Summary:
If we know a value has a tag for an interface `I` then we can only prove disjointness when comparing against a final class. In any other case involving an instance of an interface the comparison will be indeterminate.

Assume you have the following case type declared:
```
interface I1 {}

interface I2 {}

case type CT = I1 | I2;
```
Even though `I1` and `I2` do not extend each other, because of multiple inheritance of interfaces it is possible for a value to be both an `I1` and `I2`. I.e. `class C implements I1, I2 {}`.

The type checker cannot efficiently query all implementations of an interface, but even if we did if just because there doesn't exist a class that implements `I1` and `I2` today, doesn't mean there couldn't be a class introduced later down the road that can. Supporting this would allow anyone to add additional constraints of how interface `I1` and `I2` can be used in the future.

This diff makes the appropriate changes to `typing_case_types` to model this behavior.

Reviewed By: vassilmladenov

Differential Revision: D48814014

fbshipit-source-id: a7f001988f0a64c08321107e2befaec72d88e0a2