Allow Multiple Class Types in Case Types
Summary:
Building on D48768815, we now allow multiple classes to be specified in a case type. Given two classes, if neither is a subclass or another then we can conclude the classes are disjoint from one another.

This required updating the `Tag.relation` function and adding a new `Tag.class_kind` of `Class`.

We then update `DataType.Class.to_datatype` to create the tag `Tag.InstanceOf (name, Class)` when we encounter a class.

Reviewed By: vassilmladenov

Differential Revision: D48769112

fbshipit-source-id: d4f25ddcc6e73cba27cc02671609702b5cfa1cc7