Eliminate recursion / stack bloat in processMessage
Summary:
The `LogCategory::processMessage` method had ~110 bytes of objects on its stack
and recursed before those objects were destructed.  One backtrace observed five
frames of processMessage.  This is burdensome for small-stack Fibers, and
unnecessary.

If the objects were destructed earlier and the compiler was sufficiently smart,
the recursion could be inlined as a tail-call that doesn't waste a bunch of
stack each recursive step.  But the compiler cannot know that the destructors
won't need the object and so the frame can't be inlined in recursion.  We could
make the objects go out of scope before recursion, but that still requires a
sufficiently smart compiler to make the tail-call optimization.

We can save the wasted space and obviate the need for a clever compiler by just
converting the recursion into a loop.  This is almost always the better
approach.

Reviewed By: mogeb

Differential Revision: D48847992

fbshipit-source-id: 65d6af9ff8fd6ad169dd0916ccb55effc72c6362