(stresstest) some improvements regarding using io_uring
Summary:
Currently, in stress test server, set `--io-uring true` doesn't do anything.
two changes in this diff
1. if `FLAGS_io_uring` is set to true, set both `setPreferIoUring` and `setUseDefaultIoUringExecutor`

2. if `setUseDefaultIoUringExecutor` is configured, we set up `DefaultIOUringExecutor` in `configureIOUring` regardless of what is currently configured. We may want to revisit the config a bit to make it easier to set up.

For example, we may want to only use `FLAGS_io_uring` flag, and if true and `ioThreadPool_` is not set, we set the default one. Else, we check whether configured `ioThreadPool_` is compatible with io_uring. if not, we may want to crash as the server is not configured consistently rather than fall back to epoll.

Reviewed By: robertroeser

Differential Revision: D48885772

fbshipit-source-id: 768170fb6012a0a1031c8919317b4a8dcc214c45