Signature helper function
Summary:
In this diff, I create a helper function that both Completion_resolve and Completion_resolve_location can use to fetch the signature given context (ctx) and fullname.

After this diff, I will make sure that the signature is correctly updating the 'details' field in CompletionItem.
I was just having kind of an awkward time getting the signature part of 'results' in clientLsp.do_resolve_local to add to params.

Reviewed By: ljw1004

Differential Revision: D48982558

fbshipit-source-id: a8f2e6c7450848cd05a3b179a70a10a0f5039f19