Tweak order of operations in EventBase::loopMain()
Summary:
Due to how libevent handles internal events, we need to explicitly check the notification queue in the loop, but currently we do it after running the loop callbacks. This is inconsistent with what would happen when the queue is handled by libevent, and it causes the loop callback that rearms the queue to be deferred to the next loop.
This has no observable consequences currently, but complicates the invariants, so it is better for fix it.

Also, add `getNumLoopCallbacks()`, which can be useful when driving the loop externally to determine when there is any non-blocking work left to do.

Finally, better document why we need all this.

Reviewed By: Gownta

Differential Revision: D48954614

fbshipit-source-id: 57ab557029e678acfa3e1f9b52daad85528b0f52