Improve handling of functions in the parser
Summary:
Replace `return_type` with `return_clause` that contains explicit interaction (or type) name and get rid of the vector of types hack. We already decoupled sinks and streams from return types and this change completes this work on the parser level, bringing the whole thing closer to the documented grammar.

Fix a few issues in the grammar:

* Duplicate definition of `function`.
* Using `field` instead of `parameter`.
* Using `annotation` instead of `deprecated_annotations`.

Reviewed By: thedavekwon

Differential Revision: D48810369

fbshipit-source-id: 15ceb5085d829d6f9306d201ad918663bc1684dc