Strict switch: add support for noreturn types
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

This diff adds support for erroring when the case values are not a subtype of the scrutinee. It does so by adding tests which switch on an impossible value (e.g. `(int & string) = noreturn`) to trigger the new functionality. Along the way, it also tests the previously implemented (D48866898) but untested redundant default error.

It does so by (a) calling `Typing_union.simplify_unions` on the scrutinee type, which simplifies types like `(string & int)` to `noreturn` (b) partitioning the cases according to the values computed by symbolic_dnf_values. Any leftover values are ill-typed.

Reviewed By: dlreeves

Differential Revision: D48974847

fbshipit-source-id: 187dba025ca755addca5af890ff886b1f3c6bb3c