Speed up .zhhdg decompression from 37s -> 16s
Summary:
`.zhhdg` decompression got dramatically slower when the decompressor was linked into
`hh_server`, vs. being a standalone binary as I originally tested it.

I looked into it, and the problem seems to be that OCaml binaries link against
the system libc, rather than jemalloc, and Linux libc is horribly slow at cross-thread frees.
One loop in the decompressor was allocating a huge number of small edge vecs in the main
thread, then freeing all of them during a parallel algorithm from worker threads.
I think they were all getting jammed up on a free() mutex.

I didn't see this in my original testing because that was with a pure Rust `depgraph_decompress`,
which uses `jemalloc`.

To fix, I now allocate in the main thread but don't free anything in the worker threads.
I went ahead and used an arena because there is no reason to do huge numbers of small
allocations we're just going to drop at the end of the function.

This does use more memory temporarily, since the old way was freeing as it went.

A fancier version could use pipeline parallelism to overlap zunstd, delta-decoding, and file writing
to avoid that problem, but that didn't seem worth it right now.
Changing the file format to have (say) 128 independent zstd blocks that
we decompress in parallel would also shave off a few seconds.

Reviewed By: edwinsmith

Differential Revision: D48984848

fbshipit-source-id: 795b7c86e318a9226ce1511c1b6459d105732510