Load thrift standard library when present in include path
Summary: Schema generation features require having access to the AST of several thrift files that are logically coupled with the compiler, which we will call the thrift standard library. Since common usage of the thrift compiler is on files in a repo that has the standard library available we try to load the root file from the include paths. A future diff may explore bundling these files with the compiler as a fallback for usecases outside a repo.

Reviewed By: vitaut

Differential Revision: D48692692

fbshipit-source-id: 5fe9ba677463b14b93bd543ee6ed2f8d997be095