avoid persistent connection for tests
Summary:
There are four integration_ml tests here which test the prechecked algorithm. Except they test it using the persistent connection. That's not right. The prechecked algorithm is about hh_server's fanout calculations, and we should test it using stuff that hh_server actually deals with (i.e. files changing on disk), not stuff that it no longer deals with (i.e. editor files changing via the persistent connection).

This diff changes the four tests to just work with normal file-changes on disk and normal non-persistent connections to hh_server, the kind of connection that's used for e.g. "hh --find-refs".

Reviewed By: patriciamckenzie

Differential Revision: D48942615

fbshipit-source-id: ec9610787cfbda22b61d2ecb30d87a812659e349