enable logging
Summary:
This changes integration_test_ml to turn on logging.

I think logging in tests is good practice! The way buck test works is that if the test passes, then stderr is dropped. But if the test fails, then it retains stderr. Therefore, we should always have stderr just in case the test fails.

Reviewed By: nt591

Differential Revision: D48915163

fbshipit-source-id: 21b0d329258f90654bbd4038705282c0da11204a