update IDE tests from hh_server to clientIdeDaemon
Summary:
Plan of work:
1. D48938946 introduces the ability to drive clientIdeDaemon from a test
2. D48938969 (this diff) ports a load of tests from the old hh_server persistent connection over to clientIdeDaemon
3. D48938968 deletes other tests that used to make sense for hh_server persistent connection, but no longer make sense for clientIdeDaemon
4. D48939449 deletes the test-driver-code that used to drive the hh_server persistent connection

See D48938946 for overview+explanation. This diff ports a load of the tests from old hh_server persistent connection over to clientIdeDaemon.

I think this diff is largely unreviewable. Sorry! I tried to mostly do a like-for-like port, but phabricator diff-view doesn't really reflect that. I suppose there were just too many changes.

**Hh_server persistent-connection philosophy**:
1. there exists a single state of truth for every file, either the file (if it's open in the IDE) or the disk.
2. the persistent connection is able to deliver every single error for every file, and it pushes these errors when its own logic decides to.
3. there's horribly complicated logic relating to how and when things get recomputed, and decls get redefined; many of the tests illustrate that.

**clientIdeDaemon philosophy**:
1. we only ever compute errors for files open in the IDE, and those errors always reflect a truth consisting of (1) the open IDE contents for this file alone, (2) the disk content for every other file
2. errors are pulled (not pushed) in response to `open` and `edit` actions from the IDE

For many of the tests, I had to adjust the test a little so I still exercised roughly the same scenario, but I had to tweak when we open/save the file. For instance, in hh_server it's enough merely to edit an unsaved file in order for squiggles in other files to change, but in clientIdeDaemon you need to actually save a file to disk before other files change.

There's only one test which revealed oddities: `test_override` produces a Decl_elems_bug. That's because, for perf reasons, clientIdeDaemon doesn't invalidate everything that it should upon disk change. This bug will have to be fixed but it's a vast project, so I left it commented-out as a TODO.

Reviewed By: patriciamckenzie

Differential Revision: D48938969

fbshipit-source-id: a5ebcefa5491662a7ca7d576057042acf76f5ed2