move symbol_index_test off Sqlite
Summary:
(This diff continues my project to delete SqliteIndex, started+described at the base of this diffstack D47867455)

SqliteIndex is only used by tests and OSS. This diff moves one of the tests over to LocalIndex.

There's one feature which we'll lose. SqliteIndex and CustomIndex(glean) both supported a "delta" of changes on top of their databases. LocalIndex doesn't have that. I therefore just removed test coverage of the delta. That's a shame, but I can't think of any good alternatives.

Reviewed By: nt591

Differential Revision: D47869643

fbshipit-source-id: f4a49003bd7eb214f57c3fa22118960aeb83ea3d