Send package info to hh_distc
Summary:
Package information needs to be sent as part of the typecheckerOptions to hh_distc. This diff accomplishes this by defining some conversion functions in hh_config.rs.

Unfortunately, I'm having trouble using the From<> paradigm to do this because the conversion is from a type defined in hphp/hack/packages, and to a type defined in oxidized, and oxidized uses hphp/hack/packages, so the From<> conversions can't live that low level. Therefore the conversions just have to live in hh_config.rs

Differential Revision: D48971660

fbshipit-source-id: 91d20b9c31f4ba7c311ef8f93c0c64df2b781ad6