Improve memory effects of RetCtrl
Summary:
Frame context and frame func are unused after returning from the function.

This won't kill frame func and won't hurt debugging experience, because it is
stored by the EnterFrame instruction that can't be optimized out.

It should optimize away some frame context stores, but there are still
many other uses, so the effect on stores so far is minimal.

We can't do the same with AFMeta, because part of it is used for the RetCtrl
logic and other part might be used by retHelper stub.

Reviewed By: ricklavoie

Differential Revision: D48994317

fbshipit-source-id: dbaacc22f632dae295f54dbec86789bbdcf32990