Make foo.* glob foo in typechecker glob matching algo
Summary:
foo.* used to glob foo due to a bug in the glob matching algorithm in the typechecker. after the fix, we still want to keep this behavior the same.

specifically, if we have module `foo.bar`, and there are globs `foo.bar.*` and `foo.*`, we want to have `foo.bar` matched against `foo.bar.*` instead of `foo.*`

Reviewed By: jamesjwu

Differential Revision: D48985779

fbshipit-source-id: 40ccc582dc5a53914633bb9fe7a5b35dc2be57cc