allow batch-indexer to work with TestDisk
Summary:
A lot of our tests use TestDisk, which is neat because we don't need to wrap the tests in a "try-finally" which deletes leftover files after it's finished.

The rust-batch-indexer however uses `std::fs::read` to read files, which bypasses TestDisk.

This disk changes the rust-batch-indexer to make it work with TestDisk. Here were the design choices.
1. Make the rust code call back into ocaml, so it picks up either TestDisk or RealDisk appropriately. *This wouldn't work because the goal of rust is to be parallel, but ocaml has to be sequential*.
2. Shift the TestDisk infrastructure into Rust. *This would be a bunch of work that I have no appetite for*.
3. **(This diff)** Make the Rust code be able to both accept files given to it OR fetch them itself, and give files to it only in tests

Reviewed By: rjbailey

Differential Revision: D48996394

fbshipit-source-id: d13d8ec7417bc8697fec16097cdcd8974050c1bc