cleanup code for lazily indexing files
Summary:
Now that we've committed to batch-indexing of files, I can do some cleanup.

It used to be that we might accumulate changed-files-that-need-indexing at any time, either in state `During_init` (in case DidChangeWatchedFiles arrives when we're in the middle of downloading a saved-state) or in state `Initialized` (from a normal DidChangeWatchedFiles).

But now we handle the second case synchronously as soon as the message arrives.

Therefore the only place where we need to accumulate changed-files-that-need-indexing is specifically in the `During_init` state. This diff therefore moves the field `changed_files_to_process` out of common state shared by all states, and into During_init.

Reviewed By: nt591

Differential Revision: D48983188

fbshipit-source-id: b321154b3fc9a82497b0d2e6a7d5beaa27ff5d93