Issue warnings for soft deployment boundary violations in repo mode
Summary: Include both hard and soft packages in the build, so that in repo mode when attempting to call into a soft-deployed package we can raise a warning instead of fatal'ing with "undefined symbol". To this end, disable some optimizations in hhbbc when we know the deployment boundary check will fail.

Reviewed By: oulgen

Differential Revision: D48887021

fbshipit-source-id: 161f63c71c1594c76b0bac7918dfac27c3f74133