Fix pop/exception order in iopReq
Summary:
In this opcode, the interpreter is popping and decref-ing arguments from the stack before the opcode throws which violates the invariants in the interpreter.

For example, if this opcode is interpreted and we re-enter the JIT (for example using InterpOne), we would have popped an item off the stack and decref-ed it already before the throw. In the JIT, we would then end up in the catch block and eventually the unwinder which then attempts to decref these stack slots once again potentially leading to a crash.

Reviewed By: ricklavoie, jano

Differential Revision: D48890062

fbshipit-source-id: f472d8101131b43cee5769963859b672478fa53a