Refactor polymorphic variant type for missing cases into a module
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

It does so by refactoring code which uses polymorphic variants into a regular datatype inside a module to make it easier to update in a subsequent diff when we add the cases "True" and "False" for supporting boolean values.

Reviewed By: dlreeves

Differential Revision: D49062079

fbshipit-source-id: d367d305dd8777ffdf1e39a2f14e718f64f9969e