Allow client method TProcessorEventHandler callbacks to be bypassed by name
Summary:
The `EventHandlerRuntime` class is intended to provide runtime support for `EventHandlerBase` types, of which there are two of relevance to us: `TProcessorBase` and `TClientBase`.

This diff adds a mechanism to kill-switch the invocation of event callbacks on the client side based on target service or method name. In short, kill-switching on the client side is a useful mechanism to kill accidentally introduced cyclic dependencies between `TProcessorEventHandler` and low-dependency code (like a particular configuration service used at Meta).

The plan is to move other runtime support code from `EventHandlerBase` into `EventHandlerRuntime`.

`EventHandlerRuntime::isClientMethodBypassed` is called for *every* client RPC from `AsyncClient`. Therefore, read performance is fairly critical — hence the use of `folly::rcu`.

The write side will be hooked up with Thrift Flags and the performance there is not too important because updates are extremely rare. So it makes sense to squeeze out performance on the read side.

The use of `ColocatedAllocator` is to avoid page faults and improve cache locality. Even though `isClientMethodBypassed` is `O(N)` in with respect to the number of bypassed names, we expect `N` to be fairly small in practice (<10). So a linear search instead of a hashmap is not a concern.

Reviewed By: maalbash

Differential Revision: D48619419

fbshipit-source-id: 2b97dbf23ec451d544ef303082cbf226552c43f3