Pick maximal type arguments
Summary:
We were randomly selecting type arguments when dealing with union types. This was working fine for like types because the type is `union \/ t` and we choose the type arguments in the reverse order, so we always got that of `t`.

However, this doesn't work when we have `(nonnull & dynamic) | t`.

This in turn leads to incompleteness when reified type arguments are checked (e.g., `$obj->foo<int>()`), where `foo` has a reified generic so the check fired here saying it was expecting a type argument.

This diff slightly improves on this by choosing a maximal number of type arguments to put in the TAST.U

Note that the typechecker already complains if we have a union between two types with differing type parameter arity.

Differential Revision: D49055778

fbshipit-source-id: e3ca3b62fc26b65074af2cf2736e1501074fa334