Back out "Back out "[thrift-py][compiler] expand thrift_spec at import time""
Summary:
Ok so it seems that the thrift/lib/py and the compiler can not be relied on to ship at the same cadence. We had a use of the updated compiler with an older form of thrift/lib/py/Thrift.py causing an import error.  So to make sure that can't happen i'm now including the expand code directly inside the code generator.

Lets be honest this is tech debt but the continued use of thrift-py is also tech debt.  I'm not happy about this solution but the fix is more important than the purity of the thrift-py code generator.

Original commit changeset: ba138e260cc4
Original Phabricator Diff: D49087730

Reviewed By: zsol

Differential Revision: D49103162

fbshipit-source-id: 48f584efd431c0660eb86c4850f3c6977518c923