Adjust null tests
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

It does so by tweaking the names of the functions in the null-related tests to be more consistent with upcoming bool-related tests and adding an extra test for redundant default checking for null-only cases.

Reviewed By: dlreeves

Differential Revision: D49098697

fbshipit-source-id: 33c608a406f5d55c90174c75fd720ef33b76acef