Needs_writes
Summary:
If hh_server is busy doing a typecheck, and then an ide EDIT/CLOSE message comes to interrupt the typecheck, then it doesn't make sense to resume the typecheck after the interruption. We expressed this case with `Needs_writes`.

But now, with no persistent connection, we never get such an ide message, and hence the case `Needs_writes` never arises. So I'm deleting it.

Reviewed By: mheiber

Differential Revision: D49081494

fbshipit-source-id: a14716b7db7c3296dce25e2f139980b07655ef97