ServerFileSync
Summary:
It used to be that hh_server maintained data about which files were open in the IDE, and what their content was. It did this because it took those IDE content as the "source of truth" for the content of the files. It was maintained in ServerFileSync.ml.

Now with the persistent connection gone, nothing ever modifies the content of ServerFileSync any longer. So I'm just deleting it.

There used to be a few hh command-line arguments which tried to ignore IDE content, by temporarily replacing IDE content with disk content. That trick is no longer needed, so I deleted it too.

Reviewed By: mheiber

Differential Revision: D49044293

fbshipit-source-id: d6741396ac3b7b899c8192e1e14ac83b5080fe6c