Provide more information in string conversion notice for lazy classes
Summary:
Before this diff, for a function `f` that takes two string arguments, both of the following snippets

https://www.internalfb.com/code/fbsource/[87a0da194b983fbe525927ea0051b0048166e5d5]/fbcode/hphp/test/slow/class-ptr/lazyclass-string-typehint1.php?lines=8-11

https://www.internalfb.com/code/fbsource/[87a0da194b983fbe525927ea0051b0048166e5d5]/fbcode/hphp/test/slow/class-ptr/lazyclass-string-typehint2.php?lines=8-11

would produce the log line
```
Implicit Class to string conversion for type-hint in <file>.php on line 11
```

I've threaded some more information about the origin of the type being verified to the place where we emit the RaiseNotice instruction, to make the WWW codemod's job a bit simpler.

Now the former's log line will mention `argument 1 passed to f()` and the latter argument 2.

Reviewed By: arnabde03

Differential Revision: D49082826

fbshipit-source-id: e9777a00fcec251f7efaacb36485eaf9055b40d8