Introduce preg_get_error_message_if_invalid
Summary:
Developers sometimes use `preg_match($regex, '') !== false` to check if a regex is valid. However, when the regex is invalid, the runtime raises warnings for the bad regex. This diff introduces a new builtin function `preg_get_error_message_if_invalid` that does this check and returns an error message instead of raising warnings.

Will introduce the .hhi declaration in a separate, later change.

Reviewed By: oulgen

Differential Revision: D48847414

fbshipit-source-id: 9e21b41ea8591c7bfb8aba83ba61e80ef81b159a