Add support for integer literals
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

The main logic of this diff is ensuring that integer literals are not repeated. This is done by building a `Hash_tbl` from a literal of a given type (integer, bool, null) to the cases have that literal. If there is more than one case, then the last redundant case is reported as an error.

Reviewed By: dlreeves

Differential Revision: D49106845

fbshipit-source-id: beac9c83fc3acc7f2093b0e4fe3c1b9a4da6fbcf