Top Level Functions with Signatures
Summary:
Before, top level functions didn't show things like arguments/types in the autocomplete card. This made it awkward to know how to call/what to pass into these functions:
 {F1086996420}

In this diff, I fix a bug with how I'm sending over the fullname of the function (thanks ljw1004 !) and when signature is non-none (is that the way we say this in ocaml?) I add it to the detail field!

Reviewed By: mheiber

Differential Revision: D48578842

fbshipit-source-id: 5360a6eb7bda6ea77e3f083fd8a63c8cccf1067c