Refactor types used to track supported case-expressions
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

It does so by refactoring the modules and type used to represent supported literals and class constant, and using `Option.None` to represent default. The result is that the module can be used to index the hash-table in the redundancy case check, and the resulting code enforces the absence/presence of default at the type-level. This has caught previously missed cases of unsupported case-expressions.

Reviewed By: dlreeves

Differential Revision: D49190072

fbshipit-source-id: 1488664d86d7cff763d158647a65377b236b9109