Fix assert and bug
Summary:
1. There was a bug where when checking the `RenamableFunctions` allowlist, we were stripping off the entire namespace, not just the global one. Fixed the function. Added a test.

2. There was another bug where we were not casting (attrs & AttrPersistent) to a bool.

After I fixed this assert, it was firing on some generated 86cinits that were persistent (since AttrPersistent for these is just set based on systemlib), and were also renamable when `EvalJitRenameFunction=1`.

Make `isGenerated()` part of call to `funcIsRenamable()` since no generated functions are renamable.
Is there a better place for this `is_generated` helper?

Reviewed By: ricklavoie

Differential Revision: D49118702

fbshipit-source-id: df6c7b989365382bdf1c867b2b7ac43ad1b4be59