Fix a bug in op::equal that thrift struct might be incorrectly equal to itself
Summary:
In op::equal, we will first check whether two fields have the same address and short-circuiting the evaluation for such case.

The problem is that if thrift field is NAN, op::equal should always return false (even if we are comparing a thrift value to itself), since NAN are always unequal (Based on IEEE754 Spec).

This diff fixed this and added unit-test.

Reviewed By: thedavekwon

Differential Revision: D49178067

fbshipit-source-id: 0deb8dc8ffa15164bf394e864a8b64c1aaf9561b