changeResult import in simplejson_protocol to grab directly from anyhow crate instead of its own crate
Summary:
I was trying to figure out the error type of `deserialize` for some of my own code, but when I went to definition, I couldn't find out that the `Result` was an `anyhow::Result` directly.

That one extra layer of indirection seems a bit clowny. Let's get rid of it.

Reviewed By: edward-shen

Differential Revision: D49256699

fbshipit-source-id: bdf1745857bd2fb91a97acce5488e83dc8bbaab7