add support for OpenSSL 3.1.x
Summary:
Fedora 39 will ship OpenSSL 3.1.x, and at least one of the Folly tests fail
because it tests for error messages based on the detected OpenSSL variant, and
currently we can only test for OpenSSL 3.0.x and not OpenSSL 3.x.

Add `FOLLY_OPENSSL_IS_3X` that is more generic than `FOLLY_OPENSSL_IS_30X` and
use it for the SSL certificate error in `SSLErrorsTest`.

Reviewed By: Orvid

Differential Revision: D49154593

fbshipit-source-id: e8d5585bd98caf497ba35db771c2bbfa7883088a