Use nil context
Summary:
Reverting D48904560.

Compilation fails if there is an argument named "context", because it overshadows the package import.
This is not critical. Can stay like this for now, until all caller switch to ChannelClients (which requires a context variable to be passed in - so is not susceptible to the collision).

Reviewed By: awalterschulze

Differential Revision: D49301163

fbshipit-source-id: 9f23c8dcbd73460604d335263cb59acf6ba503e1