Add a per-core percentile utilization metric
Summary:
We want to replace the current host average cpu utilization with a percentile that works for a variety of hardware and network RX configurations.
The idea is that RX/RPS cores may have different utilization compared to the rest of them. This change allows tracking per-core utilization and computing a configured percentile of those utilizations.

Differential Revision: D49079695

fbshipit-source-id: ad97227fb9cedfd66eb487e78b7eb378aa27cac0