Expose Result struct
Summary:
Sadly, the legacy generator exposed (exported) "...Result" structs - which are internal implementation details.
Of course, people found ways to abuse this struct in unittests.

In order to have backwards/forwards compatibility, we need to expose this struct in the new generator as well.

In the future, we should revisit this and try to have these structs unexported.

Additionally, rename default retval name from "value" to "success".

Reviewed By: awalterschulze

Differential Revision: D49324653

fbshipit-source-id: 398a5737d388ef60ff79aa2399711f97629c8e0b