Fix staticStreamer for lazy classes and enum class labels
Summary:
staticStreamer is supposed to return output that reflects the
value as it would be written in Hack. Lazy class wasn't doing this, as
it just returned the string. Enum class labels weren't handled at all,
so we would just crash if we tried to display one. This causes HHBBC
to crash if we have tracing enabled and we encounter an enum class
label.

Reviewed By: oulgen, mdko

Differential Revision: D49356749

fbshipit-source-id: 26e61268770782633619a320de3dbca7a49579d1