Minor tweaks to BSL assign (1/2)
Summary:
We don't need to add leafs back to `toProcess()` since we are actually just going to ~~process them in this round~~ (either as a root with no parents, or selectively promoted in a bucket).

## V2
I was assuming all leaves were processed round 0. This is actually not true. Since leaves are not explicit roots, they are only processed when a parent is processed, which could be any round.
We still don't need to be adding them to `toProcess` and all of the leaf metadata computation only needs to happen the first round when we see all the leaves for the first time.

Kicked off another test

Reviewed By: ricklavoie

Differential Revision: D49024023

fbshipit-source-id: 1e45814a2a5ac215fde810da51f2c5fe55b08a4e