Make identifiers for expression IDs deterministic
Summary:
We create a new Ident_provider module. Start with reading its API and implementation.

Then we embed an Ident_provider.t in typing environments. This enforces that the ident counter is reset for each new environment, i.e. for each typing unit.

Then we use this ident provider to create all the identifiers for expression IDs.

This way, idents for expression IDs don't depend on for example the typing order for files.

Differential Revision: D49094875

fbshipit-source-id: a11f8d618506a1db08eb17e359fdeb5199eb5b47