TAST hashing: also hash the saved env
Summary:
To do well, we should also hash the saved_env in TASTs when hashing TASTs. This is what this diff does.

It omits the type inference env though because we chose to expand type variables before hashing instead.

The options are also omitted: we assume users will compare TASTs produced with the same options, making one or two flags changes.

Differential Revision: D49140791

fbshipit-source-id: 60db9117c85021002b7aefed7d2a38f50dcc968f