Support customizing UMPSCQueue segment size in SerialExecutor
Summary:
Some SerialExecutor use cases only enqueue a small number of items
into the queue at any point in time, or for the lifetime of the SerialExecutor.
For use cases which create SerialExecutors per request, for example, the
default queue segment size forces a 2^8=256 element segment by default. With
128 byte elements (for hw destructive interference reasons) this works out to
32kb of memory per segment. This is on the excessive end when maybe O(10) tasks
will be enqueued.

Reviewed By: Gownta

Differential Revision: D48412352

fbshipit-source-id: 4f9a317137d01d3360ef0d7101af02b03048f12d