Unset all temps in the non-exception case too
Summary: We take care to unset all temporary variables when leaving their scope (as determined by the `with_unnamed_locals` function in the emitter) via `Throw`. However, we only unset some of them when leaving the scope via normal control flow. This is because the unset bytecodes for the exception path are calculated to be all of the temporaries allocated during the `with_unnamed_locals`, but the unset bytecodes for normal control flow are computed separately at each use of `with_unnamed_locals` and some of these miss e.g., temporaries used in a sub-expression.

Reviewed By: aorenste

Differential Revision: D49325282

fbshipit-source-id: 231921f93f7bb8ef700a4b71413611b54787b148